fix(engine): expand SSR verification batch hash coverage, fix i64 overflow in validation receipt, and enhance module interop gate scripts
react_ssr_verification.rs:
- Include overall_verdict, total_divergence_score, verified_count,
  divergent_count, and inconclusive_count in the verify_batch()
  content hash, ensuring the hash captures the full batch summary
  rather than just schema version and per-result verdicts.

translation_validation_receipt.rs:
- Widen proven_rate_millionths intermediate product from i64 to i128
  to prevent overflow when total_proven is large (same pattern as
  the interleaving_explorer fix in the previous commit).

module_compatibility_matrix.rs:
- Add assertions for rch artifact-retrieval and timeout failure path
  markers in the module interop gate script.

Scripts enhanced:
- rgc_module_resolution_trace_contract_smoke.sh: expanded validation.
- run_rgc_module_interop_verification_matrix.sh: additional checks.

Cargo.lock updated with dependency version changes.

+++ crates/franken-engine/src/translation_validation_receipt.rs
@@ -1140,7 +1140,7 @@ impl ValidationReceiptEmitter {
     pub fn summary(&self) -> ReceiptSummary {
         let chain_integrity = self.chain.verify_integrity();
         let proven_rate_millionths = if self.stats.total_receipts > 0 {
-            (self.stats.total_proven as i64 * 1_000_000) / self.stats.total_receipts as i64
+            (self.stats.total_proven as i128 * 1_000_000 / self.stats.total_receipts as i128) as i64
         } else {
             0
         };
